remove usage of outdated pyschemes
