+++ cg/meta/orders/__init__.py
@@ -1,2 +1,1 @@
 from .api import OrdersAPI
-from .schema import OrderType
